Create redis instance builder
* Move logger builder to separate lib file
