+++ workshop014-cache/src/setup/logger.js
@@ -1,17 +1,5 @@
-const pino = require("pino");
+const pinoLogger = require("../lib/pinoLogger");
 
 module.exports = function (app) {
-  const pinoOptions =
-    process.env.NODE_ENV !== "production"
-      ? {
-          transport: {
-            target: "pino-pretty",
-            options: {
-              colorize: true,
-            },
-          },
-        }
-      : undefined;
-
-  app.locals.logger = pino({ level: process.env.LOG_LEVEL, ...pinoOptions });
+  app.locals.logger = pinoLogger;
 };
